autoparser: add schema-path option, use old-style container types
Instead of specifying schemas as URLs, which would have added a
dependency on requests as well as requiring outbound network
connections, add a --schema-path option to both scripts that allows one
to specify a folder where the schemas are located.

Also use old-style container types (from typing module) to have compat
with Python 3.8.