SDTM MB domain to FHIR mapping
Initial thoughts, mapping not completed. IPS = International Patient Summary and forms an ISO standard for data sharing. Pathogen testing is included as a resource in the IPS specification so I think it would be good to ensure we align with this where possible. Commented in the mapping sheet where IPS-aligned terminology is used